Add HTTY transport support and refactor environment layer
- Add `Async::HTTY` transport support: when `ENV["HTTY"] == "1"`, the
  server runs over a terminal side-channel instead of a TCP socket.

- Refactor environment modules into focused layers:
  - `Lively::Environment::Middleware` — shared application/asset stack
  - `Lively::Environment::HTTP` — Falcon TCP transport + Middleware
  - `Lively::Environment::HTTY` — HTTY transport + Middleware
  - `Lively::Environment::Application` — multiplexes via `make_service`,
    using the new `Async::Service::Generic#make_service` hook (requires
    async-service ~> 0.23)

- Force `Async::Container::Threaded` in `bin/lively` — no forking needed
  for a single-user dev tool.

- Fix `Live.js` WebSocket URL construction to handle `htty:` scheme,
  mapping it to `ws:` alongside the existing `http:` → `ws:` mapping.

- Fix `Lively::Assets` missing `require "uri"` (was implicit via Falcon).
