+++ lib/lively/environment/application.rb
@@ -1,56 +1,68 @@
 # frozen_string_literal: true
 
 # Released under the MIT License.
-# Copyright, 2021-2025, by Samuel Williams.
+# Copyright, 2021-2026, by Samuel Williams.
 
-require_relative "../application"
-require_relative "../assets"
-
-require "falcon/environment/server"
+require_relative "middleware"
+require_relative "http"
+require_relative "htty"
 
 # @namespace
 module Lively
 	# @namespace
 	module Environment
-		# Represents the environment configuration for a Lively application server.
-		# 
-		# This module provides server configuration including URL binding, process count,
-		# application class resolution, and middleware stack setup. It integrates with
-		# Falcon's server environment to provide a complete hosting solution.
+		# Multiplexing environment for Lively applications.
+		#
+		# Declares the transport selection as explicit, overridable evaluator keys and uses `make_service` to compose the appropriate child environment at service startup time. This keeps transport selection in the service layer rather than in module inclusion hooks.
+		#
+		# The `htty` key controls which transport is used. Override it in a service block to force a specific transport regardless of the environment variable:
+		#
+		# ~~~ ruby
+		# service "myapp" do
+		#   include Lively::Environment::Application
+		#   def htty = false  # always use HTTP
+		# end
+		# ~~~
 		module Application
-			include Falcon::Environment::Server
+			include Lively::Environment::Middleware
+			# Note: does not include Falcon::Environment::Server directly. Falcon is
+			# brought in exclusively via http_environment so that the combined
+			# evaluator's service_class resolves correctly without shadowing.
 			
-			# Get the server URL for this application.
-			# @returns [String] The base URL where the server will be accessible.
-			def url
-				"http://localhost:9292"
+			# Whether to use HTTY transport. Reads ENV["HTTY"] by default.
+			# @returns [Boolean]
+			def htty
+				ENV["HTTY"] == "1"
 			end
 			
-			# Get the number of server processes to run.
-			# @returns [Integer] The number of worker processes.
-			def count
-				1
+			# The environment module to use for HTTY transport.
+			# @returns [Module]
+			def htty_environment
+				Lively::Environment::HTTY
 			end
 			
-			# Resolve the application class to use.
-			# @returns [Class] The application class, either user-defined or default.
-			def application
-				if Object.const_defined?(:Application)
-					Object.const_get(:Application)
-				else
-					Console.warn(self, "No Application class defined, using default.")
-					::Lively::Application
-				end
+			# The environment module to use for HTTP transport.
+			# @returns [Module]
+			def http_environment
+				Lively::Environment::HTTP
 			end
 			
-			# Build the middleware stack for this application.
-			# @returns [Protocol::HTTP::Middleware] The complete middleware stack.
-			def middleware
-				::Protocol::HTTP::Middleware.build do |builder|
-					builder.use Lively::Assets, root: File.expand_path("public", self.root)
-					builder.use Lively::Assets, root: File.expand_path("../../../public", __dir__)
-					builder.use self.application
-				end
+			# The environment module for the selected transport.
+			# @returns [Module]
+			def transport_environment
+				htty ? htty_environment : http_environment
+			end
+			
+			# Build the service by composing the transport environment on top of this one.
+			# Called by Async::Service::Generic.wrap — self is the evaluator at call time.
+			# @parameter environment [Async::Service::Environment]
+			# @returns [Async::Service::Generic]
+			def make_service(environment)
+				combined = environment.with(transport_environment)
+				combined_evaluator = combined.evaluator
+
+				# Call `service_class.new` directly rather than `Async::Service::Generic.wrap` — the combined evaluator still has `Application` (and therefore `make_service`) in its ancestor chain, so `wrap` would recurse back into this method.
+				return combined_evaluator.service_class.new(combined, combined_evaluator)
 			end
 		end
 	end
