+++ lively.gemspec
@@ -27,6 +27,8 @@ Gem::Specification.new do |spec|
 	spec.required_ruby_version = ">= 3.3"
 	
 	spec.add_dependency "agent-context"
+	spec.add_dependency "async-htty"
+	spec.add_dependency "async-service", "~> 0.23"
 	spec.add_dependency "falcon", "~> 0.47"
 	spec.add_dependency "io-watch"
 	spec.add_dependency "live", "~> 0.18"
